Added support for topojson and data resources in non-json format.
Now the topojson module is loaded when the WITH_TOPOJSON define 
is declared, moreover the d3.geo.projection extension is loaded 
when the WITH_JSLIBS define is declared.

Moreover a new customized version of d3 including the dsv module 
allows to load data resources in a format different from json.